+++ scripts/utils.py
@@ -1,0 +1,3 @@
+import pathlib
+
+REPO_ROOT = str(pathlib.Path(__file__).resolve().parent.parent)